bomberman/requirements.txt
pygame==2.5.2
mysql-connector-python==8.2.0
python-socketio==5.10.0
python-socketio[client]==5.10.0
aiohttp==3.9.1